Updated Cutscenes
Cutscene for leaving the house has been updated to go through the door. The cutscene for exiting level one has been created and will fade to black as it is finishing. The level switch from level one to level two needs to be re added to trigger when the cutscene finishes