Changes To Follow Spirit BT and added spirit that moves away
Changed the behaviour tree of spirit that follows player to go back to its spawn and stay there if player is not within range. Added spirit that will stay stationary (next to one of the events) and will move away when player gets close. Will add in they will only move if interacted with if required.