Added Spirits into Map and Made Spirit Changes
Spirits were added to the locations on the map, can be changed for final build. Changed spirit Aid so it will always face the direction of the player. The AI roam has been added in as well so it won't collide with player. The EQS AI will now roam about it's area when the player is not within it's radius to appear more lifelike.